Add Mod Export Options
In Render->Armory Project->Modding you can now choose a modding mode
for the game. Setting the mode to "Game" will cause shaders to be loaded
from the data directory as opposed to being compiled into generated
code. Setting the mode to "Mod" will allow you to export the game built
as a mod for another Armory game.