Fix for faulty "missing uv maps"
Ask for export_uvs to the original object instead of the evaluated copy. Reason is, it seems the dependency graph fails to update
whenever custom properties are set. So for that I think asking the original object in this case could be sufficient to fix it.

The same could be applied to the other functions that are in that chunk of code.