Fixed memory problem in `ArmoryExporter.execute` [fixes #1604]
The Blender crash referenced in the issue happened always
when starting 2nd iteration of the loop in line 2041. I'm not
really sure, but I bet it had to do with
`export_object` modifying `bpy.context.collection.objects`
(line 818), wich in turn could invalidate the reference
obtained in line 1926, if I understand a bit of Blender
internals — wich I didn't some days ago, so this could be
completely wrong.

It no longer happens with this change.