Fix rigid bodies continuing to move when parented
This fixes one more bug from here: https://github.com/armory3d/armory/issues/1830

I don't know if this is a correct fix, but solved a problem for me.

One last bug that stills is the rigid body in "random" places when parented. To fix this, when the rigid body (no matter if active or passive) is parented, it must have Animated checkbox checked. I don't know how to do this. This will solve all problems related to it, except the "Keep Transform" unchecked one.