Fix and improve loading of libraries
Fixes an issue described in https://forums.armory3d.org/t/celshade-option-only-appears-after-opening-forest-example-with-right-click-how-to-set-manually-for-other-blends/4891.

Libraries weren't reloaded when changing files from within the same project, but registered render path drivers are removed each time a file is changed. Now libraries are reloaded each time a file is opened or the addon is reloaded, which also gives libraries the possibility to re-initialize itself for a new file. Also, libraries can now have an optional `unregister` function that will be called before a new file is loaded or when the addon is unregistered, giving libraries the possibility to properly clean up.

Furthermore libraries no longer aren't loaded twice when a blend file is opened from the file explorer and there is a little debug message in the console that shows which libraries get (un-)loaded.