Fix node tree caching
Fixes https://github.com/armory3d/armory/issues/2219.

Since https://github.com/armory3d/armory/commit/7a1acb6228ecbb76013e57973c50d70270182275 there was a workaround in place to ensure that the cache of logic trees gets cleared very frequently because the depsgraph updates don't reflect all updates to logic trees. But because the `always()` callback only runs twice a second, it could happen that changes made to a tree weren't reflected in the build if the user starts the build before the `always()` callback was run. Now that we have dedicated tree update callbacks for live patching, we can use them to clear the cache on each change made to the node tree and get rid of the old workaround.

Note that the current approach is still a bit over-conservative (less than before though), for example the cache is also cleared when you start dragging a connection from a socket but cancel the operation before reaching the target socket. So there is still some potential for improvement in the future.