Fix color ramp node for factors left of the left-most stop
Fixes another issue mentioned in https://github.com/armory3d/armory/issues/2467 (see https://forums.armory3d.org/t/textured-terrain-rgb-channel-flipped/4850).

In the color ramp node it could happen that the input factor was left of the left-most stop and thus the relative position between two stops was negative. As a consequence the output of `mix(stopLeft, stopRight, relativePos)` was wrong. Because there is always an implicit stop at position 1.0, the same issue doesn't happen for the opposite case of values right to the right-most stop.

I also made the code a bit more readable to better describe what it does.