Fix numerical robustness issue for packFloatInt16 in gbuffer.glsl
In the previous implementation, a matid of 2 and a metallic value of 0.0 would result in a packed float value of 0.125, which then got unpacked into matid = 1 and metallic = 1.0. The new implementation does not have this issue, although the principle is basically the same.

Also removed the packFloatInt8 implementation which was not used (anymore?)